my generator32 init

+++ generator/graph_generator.h
@@ -49,21 +49,21 @@ static inline void write_edge(packed_edge* p, int64_t v0, int64_t v1) {
 #else
 
 typedef struct packed_edge {
-  int64_t v0;
-  int64_t v1;
+  int32_t v0;
+  int32_t v1;
 } packed_edge;
 
-static inline int64_t get_v0_from_edge(const packed_edge* p) {
+static inline int32_t get_v0_from_edge(const packed_edge* p) {
   return p->v0;
 }
 
-static inline int64_t get_v1_from_edge(const packed_edge* p) {
+static inline int32_t get_v1_from_edge(const packed_edge* p) {
   return p->v1;
 }
 
 static inline void write_edge(packed_edge* p, int64_t v0, int64_t v1) {
-  p->v0 = v0;
-  p->v1 = v1;
+  p->v0 = (uint32_t)v0;
+  p->v1 = (uint32_t)v1;
 }
 
 #endif
